refactor: extract functionality to different files
